clean data
I subset data to Baltimore, MD, converted reported_date to a date class, extracted month information to put in a new column, and created a season column to distinguish warm (Summer) and cold (Winter) months. Winter and Summer were used because these were the values in the legend shown on the homework plot.