updated value for mapping ref line and label
After closer inspection of the model plot, it looks like the ref line was mapping to the beginning of the month. I assumed this was a result of creating a date with a supplied year and month (as I did for the homicides_baltimore data frame). Since I want to match the model plot as closely as possible, I created a date this way and used it for positioning the reference line and label. I also increased the fig width to better match the model.